Remove coupling between listener and repo
This commit contains 3 main changes
1. Remove GetListener from IRepository and any derived classes
2. Remove inheritance from ICatchupStreamSubscriber out of EventStoreRepository and create a separate EventStoreCatchupStreamSubscriber that can be passed into a Listener. N.B: MockEventStoreRepository has not been changed as part of this. The reason is this repository seems to be used as both repository and subscriber/listener at the same time, and that MIGHT be acceptable for unit test, though still I think separating those must be better.
3. Fix the unit tests touched by those changes

(RD-20) WIP
(RD-32) Fixed
