+++ src/ReactiveDomain.Foundation.Tests/EventStore/MockEventStoreRepository.cs
@@ -235,10 +235,6 @@ namespace ReactiveDomain.Foundation.Tests.EventStore
                     bus.Publish(tupple.Item2);
             }
         }
-        public IListener GetListener(string name, bool sync = false)
-        {
-            return new SynchronizableStreamListener(name, this, sync);
-        }
 
         public IDisposable SubscribeToStreamFrom(
                                                          string stream,
