+++ src/ReactiveDomain.Foundation.Tests/Logging/when_logging_disabled_and_events_are_published.cs
@@ -1,9 +1,9 @@
-﻿using System;
-using ReactiveDomain.Foundation.Tests.EventStore;
+﻿using ReactiveDomain.Foundation.EventStore;
 using ReactiveDomain.Messaging;
 using ReactiveDomain.Messaging.Bus;
 using ReactiveDomain.Messaging.Testing;
 using ReactiveDomain.Testing;
+using System;
 using Xunit;
 using Xunit.Sdk;
 
@@ -35,7 +35,7 @@ namespace ReactiveDomain.Foundation.Tests.Logging
         protected override void When()
         {
 
-            _listener = Repo.GetListener(Logging.FullStreamName);
+            _listener = new SynchronizableStreamListener(Logging.FullStreamName, Subscriber, StreamNameBuilder);
             _listener.EventStream.Subscribe<DomainEvent>(this);
 
             _listener.Start(Logging.FullStreamName);
