+++ src/ReactiveDomain.Foundation.Tests/Logging/when_logging_high_volume_message_traffic.cs
@@ -1,10 +1,9 @@
-﻿using System;
-using ReactiveDomain.Foundation.EventStore;
-using ReactiveDomain.Foundation.Tests.EventStore;
+﻿using ReactiveDomain.Foundation.EventStore;
 using ReactiveDomain.Messaging;
 using ReactiveDomain.Messaging.Bus;
 using ReactiveDomain.Messaging.Testing;
 using ReactiveDomain.Testing;
+using System;
 using Xunit;
 
 namespace ReactiveDomain.Foundation.Tests.Logging
@@ -48,7 +47,7 @@ namespace ReactiveDomain.Foundation.Tests.Logging
              _numberOfItemsLogged = 0;
              _catchupSubscriptionMsgs = 0;
 
-            _listener = Repo.GetListener(Logging.FullStreamName);
+            _listener = new SynchronizableStreamListener(Logging.FullStreamName, Subscriber, StreamNameBuilder);
             _listener.EventStream.Subscribe<Message>(this);
 
             _listener.Start(Logging.FullStreamName);
